add test coverage for all new components
